improve demo

+++ main/sdxl/sdxl_text_encoder.py
@@ -2,7 +2,7 @@ from main.utils import import_model_class_from_model_name_or_path
 import torch 
 
 class SDXLTextEncoder(torch.nn.Module):
-    def __init__(self, args, accelerator) -> None:
+    def __init__(self, args, accelerator, dtype=torch.float32) -> None:
         super().__init__()
         text_encoder_cls_one = import_model_class_from_model_name_or_path(
             args.model_id, args.revision
@@ -13,11 +13,11 @@ class SDXLTextEncoder(torch.nn.Module):
 
         self.text_encoder_one = text_encoder_cls_one.from_pretrained(
             args.model_id, subfolder="text_encoder", revision=args.revision
-        ).to(accelerator.device).float()
+        ).to(accelerator.device).to(dtype=dtype)
 
         self.text_encoder_two = text_encoder_cls_two.from_pretrained(
             args.model_id, subfolder="text_encoder_2", revision=args.revision
-        ).to(accelerator.device).float()
+        ).to(accelerator.device).to(dtype=dtype)
 
         self.accelerator = accelerator
